Support for the DOS II+ (and XDOS) file management system directory flag idiosyncrasy.
When a disk is formatted to medium density (130 KB), DOS II+ sets the $42 directory flags
for all files, and not only files located belov sector 720. This has been observed by formatting a blank disk to medium density with DOS II+ and adding files until the disk was full.

This commit adds one parameter to the constructors of the dos2 and expanded_vtoc classes.
The parameter is force_dos2_flag with default value of false.
When the parameter is true, files located above sector 720 do not have their DOS 2 flag reset.

The dos_IIplus constructor (and therefore XDOS) calls the super class constructor with force_dos2_flag=true, so files of the XDOS and DOS II+ filesystems always have $42 in their directory entries. Other file systems DOS2 DOS 2.5 etc. are not affected.